DoTrade Rewrite
Rewrote function to make it easier to understand. I stopped trying to
figure out ever little detail that was done in the last function and
wrote the item swap in a new way.

Restrictions Added:
1) Can't trade equip slot items (this might impact /take command)
2) Can't trade soulbound items

Attempted fixes:
1) No more items in slots they shouldn't be in.

In the event that the user trades more than the other player can carry,
the items are simply lost. A message is displayed to the user that this
happened. This should only happen when someone trys to exploit trading
in some why that results in this event.