test projectile memory leak patch
This one is kind of odd. There is a dictionary in the projectile class
that saves when an object is hit with projectiles that
passcover/multihit. So it logs this info on what the projectile hits
except it doesn't seem to be used by anything. Just there collecting
objects.

Another odd thing is why the projectile class isn't gc'ed when it falls
out of scope of the projectile[] array of the entity.

In any case it appears as if gc properly cleans up these projectile
objects with the removal of the hitted variable.